stock-prediction-platform/services/kafka-consumer/requirements.txt
confluent-kafka==2.4.0
psycopg2-binary==2.9.9
tenacity==8.3.0
pydantic-settings==2.2.1
structlog==24.2.0
pytest==8.2.0
